Validate checksum of each segment file post download from remote store
