+++ server/src/main/java/org/opensearch/index/store/RemoteSegmentStoreDirectory.java
@@ -297,6 +297,10 @@ public final class RemoteSegmentStoreDirectory extends FilterDirectory implement
                 );
             }
         }
+
+        public int getWrittenByMajor() {
+            return writtenByMajor;
+        }
     }
 
     /**
